chore(frontend): reorganizar chakra UI
Modifica a forma que o tema customizado é aplicado ao Chakra UI
(adicionando diretamente aos seus componentes ao invés de criar
novos com base nos do Chakra UI), além de reorganizar a pasta
components e remover componentes inutilizados.
