--- Front/src/components/ui/button.tsx
@@ -1,4 +1,0 @@
-import { chakra } from "@chakra-ui/react";
-import { recipes } from "../../styles/recipes";
-
-export const Button = chakra("button", recipes.buttonRecipe)
